ppo/requirements.txt
absl-py==0.13.0
ansible==2.9.27
astroid==2.3.3
attrs==21.2.0
Babel==2.5.1
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bcc==0.19.0
bcrypt==3.1.6
beautifulsoup4==4.6.3
blivet==3.4.0
Brlapi==0.6.7
cached-property==1.5.2
cachetools==4.2.2
cattrs==1.0.0
certifi==2021.5.30
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.3
chrome-gnome-shell==0.0.0
cloudpickle==1.6.0
conda==4.10.3
conda-package-handling==1.7.3
configobj==5.0.6
configshell-fb==1.1.28
cpplint==1.5.5
cryptography==3.2.1
cssselect==0.9.2
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.11.0
dasbus==1.2
dataclasses==0.8
dbus-python==1.2.4
decorator==4.2.1
distlib==0.3.2
distro==1.4.0
docutils==0.14
ethtool==0.14
fail2ban==0.11.2
file-magic==0.3.0
filelock==3.0.12
frozendict==1.2
future==0.18.2
Glances==3.1.4.1
google-auth==1.33.1
google-auth-oauthlib==0.4.4
gpg==1.13.1
grpcio==1.38.1
h5py==3.1.0
html5lib==0.999999999
humanize==0.5.1
idna==2.5
importlib-metadata==4.6.1
importlib-resources==5.2.0
iniparse==0.4
initial-setup==0.3.81.7
iotop==0.6
ipython==7.16.1
ipython-genutils==0.1.0
isc==2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.0
jsonpointer==1.10
kiwisolver==1.3.1
kmod==0.1
langtable==0.0.51
lazy-object-proxy==1.4.3
libcomps==0.1.16
libvirt-python==6.0.0
lit==12.0.1
lxml==4.2.3
Mako==1.0.6.dev0
Markdown==3.3.4
MarkupSafe==0.23
matplotlib==3.3.4
mccabe==0.6.1
# Editable install with no version control (mlagents==0.27.0)
-e /dcs/18/u1830744/diss/venv/test-env2/lib/python3.6/site-packages/mlagents-0.27.0-py3.6.egg
# Editable install with no version control (mlagents-envs==0.27.0)
-e /dcs/18/u1830744/diss/venv/test-env2/lib/python3.6/site-packages/mlagents_envs-0.27.0-py3.6.egg
netaddr==0.7.19
nftables==0.1
ntplib==0.3.3
numpy==1.19.5
oauthlib==3.1.1
ordered-set==2.0.2
paramiko==2.4.3
parso==0.5.1
perf==0.1
pexpect==4.3.1
pickleshare==0.7.5
pid==2.1.1
Pillow==8.3.1
platformdirs==2.0.2
ply==3.9
productmd==1.11
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.4.3
psycopg2==2.7.5
ptyprocess==0.5.2
pwquality==1.4.4
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.3
pycosat==0.6.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
Pygments==2.2.0
pygobject==3.28.3
pyinotify==0.9.6
pykickstart==3.16.14
pylint==2.4.4
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyparted==3.11.7
PySocks==1.6.8
python-augeas==0.5.0
python-dateutil==2.8.2
python-dmidecode==3.12.2
python-linux-procfs==0.6.3
python-meh==0.47.2
pytz==2017.2
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.26.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rope==0.14.0
rpm==4.14.3
rsa==4.7.2
rtslib-fb==2.1.74
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.3.0
simpleline==1.1.1
six==1.13.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
SSSDConfig==2.5.2
subscription-manager==1.28.21
syspurpose==1.28.21
system-storage-manager==1.4
systemd-python==234
targetcli-fb==2.1.53
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toolz==0.10.0
torch==1.8.1
tqdm==4.50.2
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.10.0.0
urllib3==1.24.2
urwid==1.3.1
uuid==1.30
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.11.2
zipp==3.5.0
